Allow backing list to shrink and reclaim tiles
If you want to change your list and shrink the number of items, this
stops the old 'extra' tiles being drawn. New tiles are created as normal
if the list grows again. You do have to call virtualize again when the
list gets smaller as theres no observer on it's size